group some fields in a common struct
