+++ tests/stop_point_discovery_test.rs
@@ -25,8 +25,8 @@ fn sp_discovery_integration_test() {
 
     let resp: SiriResponse = serde_json::from_str(body).unwrap();
     let spd = resp.siri.stop_points_delivery.unwrap();
-    assert_eq!(spd.version, "2.0");
-    assert_eq!(spd.status, true);
+    assert_eq!(spd.common.version, "2.0");
+    assert_eq!(spd.common.status, Some(true));
     // no filtering, we fetch all stops
     assert_eq!(spd.annotated_stop_point.len(), 1);
